% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simfdata.R
\name{expbh}
\alias{expbh}
\title{Exponential baseline hazard}
\usage{
expbh(t, shape = 2)
}
\arguments{
\item{t}{time}

\item{shape}{shape parameter}
}
\value{
hazard function value under Exponential distibution
}
\description{
Exponential baseline hazard
}
