% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simfdata.R
\name{gompbh}
\alias{gompbh}
\title{Gompartz baseline hazard}
\usage{
gompbh(t, shape = 2, scale = 1)
}
\arguments{
\item{t}{time}

\item{shape}{shape parameter}

\item{scale}{scale parameter}
}
\value{
hazard function value under Gompartz distibution
}
\description{
Gompartz baseline hazard
}
