% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.cphGM}
\alias{print.cphGM}
\title{print function for cphGM}
\usage{
\method{print}{cphGM}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{others}
}
\value{
prints table containing various parameter estimates,
        SE, P-value.
}
\description{
S3 print method for class 'cphGM'
}
\examples{
\donttest{
##
n1<-1000
p1<-2
X1<-matrix(rnorm(n1*p1),n1,p1)
simulated_data<-simfdata(n=1000,beta=c(0.5,0.5),fvar=0.5,X=X1)
model1<-cphGM(formula=Surv(time,status)~X1+X2,
fterm=c('gamma','id'),Time="time",status="status",
id="id",data=simulated_data,bhdist='weibull')
print(model1)
##
}
}
