% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{simulated_data}
\alias{simulated_data}
\title{Simulated multistate data}
\format{
a tibble of 13 columns and 2204 observations,
\describe{
\item{id}{id value for subjects}
\item{status}{survival status}
\item{time}{survival time}
\item{x1}{Numeric covariate}
\item{x2}{Numeric covariate}
\item{x3}{Numeric covariate}
\item{x4}{Numeric covariate}
}
}
\usage{
data(simulated_data)
}
\description{
A simulated multi state dataset used for demonstration purpose.
}
\references{
Vishwakarma, G. K., Bhattacherjee, A., Rajbongshi, B. K., & Tripathy, A. (2024). Censored imputation of time to event outcome through survival proximity score method. Journal of Computational and Applied Mathematics, 116103;

Bhattacharjee, A., Vishwakarma, G. K., Tripathy, A., & Rajbongshi, B. K. (2024). Competing risk multistate censored data modeling by propensity score matching method. Scientific Reports, 14(1), 4368.
}
\keyword{datasets}
