% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simfdata.R
\name{weibulbh}
\alias{weibulbh}
\title{Weibull baseline hazard}
\usage{
weibulbh(t, shape = 2, scale = 1)
}
\arguments{
\item{t}{time}

\item{shape}{shape parameter}

\item{scale}{scale parameter}
}
\value{
hazard function value under Weibull distibution
}
\description{
Weibull baseline hazard
}
